Move AesEax definition to aes.h.
